+++ soroban-sdk/src/deploy.rs
@@ -1,9 +1,90 @@
-use crate::{
-    env::internal::Env as _,
-    namespace::{CurrentNamespace, Ed25519Namespace},
-    Bytes, BytesN, Env, TryFromVal,
-};
+//! Deploy provides types for deploying contracts.
+//!
+//! Contracts are assigned an ID that is derived from the namespace that the
+//! contract was deployed with.
+//!
+//! - [CurrentNamespace] – A contract deployed by the currently executing
+//! contract with [CurrentNamespace] will have an ID derived from the currently
+//! executing contract's ID.
+//!
+//! - [ContractNamespace] – A contract deployed by some other contract will have
+//! an ID derived from its contract ID.
+//!
+//! - [Ed25519Namespace] – A contract deployed by the currently executing
+//! contract with the [Ed25519Namespace] will have an ID derived from its
+//! ed25519 public key.
+//!
+//! The deployer for a namespace can be created using [Env::deployer].
+//!
+//! The contract ID for a deployed contract can be derived using [Env::contract_id].
 
+use crate::{env::internal::Env as _, Bytes, BytesN, Env, TryFromVal};
+
+/// Namespace defines the namespace that other contracts IDs are defined with.
+pub enum Namespace {
+    Current(CurrentNamespace),
+    Contract(ContractNamespace),
+    Ed25519(Ed25519Namespace),
+}
+
+/// Namespace of the currently executing contract.
+#[derive(Debug, Copy, Clone, PartialEq, PartialOrd)]
+pub struct CurrentNamespace;
+
+/// Namespace of a specific contract ID.
+#[derive(Debug, Clone, PartialEq, PartialOrd)]
+pub struct ContractNamespace {
+    pub contract_id: BytesN<32>,
+}
+
+/// Namespace of an ed25519 public key.
+#[derive(Debug, Clone, PartialEq, PartialOrd)]
+pub struct Ed25519Namespace {
+    pub public_key: BytesN<32>,
+}
+
+impl From<CurrentNamespace> for Namespace {
+    fn from(v: CurrentNamespace) -> Self {
+        Self::Current(v)
+    }
+}
+
+impl From<&CurrentNamespace> for Namespace {
+    fn from(v: &CurrentNamespace) -> Self {
+        Self::Current(*v)
+    }
+}
+
+impl From<ContractNamespace> for Namespace {
+    fn from(v: ContractNamespace) -> Self {
+        Self::Contract(v)
+    }
+}
+
+impl From<&ContractNamespace> for Namespace {
+    fn from(v: &ContractNamespace) -> Self {
+        Self::Contract(v.clone())
+    }
+}
+
+impl From<Ed25519Namespace> for Namespace {
+    fn from(v: Ed25519Namespace) -> Self {
+        Self::Ed25519(v)
+    }
+}
+
+impl From<&Ed25519Namespace> for Namespace {
+    fn from(v: &Ed25519Namespace) -> Self {
+        Self::Ed25519(v.clone())
+    }
+}
+
+/// Deployer is a [Namespace] that the currently executing contract can deploy
+/// contracts for.
+///
+/// The currently executing contract can deploy contracts with two namespaces:
+/// - [CurrentNamespace]
+/// - [Ed25519Namespace]
 pub trait Deployer<D> {
     fn deployer(&self, env: &Env) -> D;
 }
@@ -38,6 +119,8 @@ impl Deployer<Ed25519NamespaceDeployer> for &Ed25519Namespace {
     }
 }
 
+/// Provides functionality for deploying a contract with a contract ID derived
+/// from the [CurrentNamespace].
 pub struct CurrentNamespaceDeployer {
     env: Env,
     namespace: CurrentNamespace,
@@ -51,13 +134,25 @@ impl CurrentNamespaceDeployer {
         }
     }
 
-    pub fn deploy_wasm(&self, salt: impl Into<Bytes>, wasm: impl Into<Bytes>) -> BytesN<32> {
+    /// Deploy a contract.
+    ///
+    /// The contract ID from the currently executing contract and the given salt
+    /// will be used to derive a contract ID for the deployed contract.
+    ///
+    /// Returns the deployed contract's ID.
+    pub fn deploy(&self, salt: impl Into<Bytes>, wasm: impl Into<Bytes>) -> BytesN<32> {
         let env = &self.env;
         let id =
             env.create_contract_from_contract(wasm.into().to_object(), salt.into().to_object());
         BytesN::<32>::try_from_val(env, id).unwrap()
     }
 
+    /// Deploy a built-in token contract.
+    ///
+    /// The contract ID from the currently executing contract and the given salt
+    /// will be used to derive a contract ID for the deployed contract.
+    ///
+    /// Returns the deployed contract's ID.
     pub fn deploy_token(&self, salt: impl Into<Bytes>) -> BytesN<32> {
         let env = &self.env;
         let id = env.create_token_from_contract(salt.into().to_object());
@@ -65,6 +160,8 @@ impl CurrentNamespaceDeployer {
     }
 }
 
+/// Provides functionality for deploying a contract with a contract ID derived
+/// from a [Ed25519Namespace].
 pub struct Ed25519NamespaceDeployer {
     env: Env,
     namespace: Ed25519Namespace,
@@ -78,7 +175,13 @@ impl Ed25519NamespaceDeployer {
         }
     }
 
-    pub fn deploy_wasm(
+    /// Deploy a contract.
+    ///
+    /// The ed25519 public key and the given salt will be used to derive a
+    /// contract ID for the deployed contract.
+    ///
+    /// Returns the deployed contract's ID.
+    pub fn deploy(
         &self,
         salt: impl Into<Bytes>,
         wasm: impl Into<Bytes>,
@@ -94,6 +197,12 @@ impl Ed25519NamespaceDeployer {
         BytesN::<32>::try_from_val(env, id).unwrap()
     }
 
+    /// Deploy a built-in token contract.
+    ///
+    /// The ed25519 public key and the given salt will be used to derive a
+    /// contract ID for the deployed contract.
+    ///
+    /// Returns the deployed contract's ID.
     pub fn deploy_token(
         &self,
         salt: impl Into<Bytes>,
