+++ soroban-sdk/src/env.rs
@@ -47,7 +47,13 @@ pub type EnvType<V> = internal::EnvVal<Env, V>;
 pub type EnvVal = internal::EnvVal<Env, RawVal>;
 pub type EnvObj = internal::EnvVal<Env, Object>;
 
-use crate::{deploy::Deployer, Bytes, BytesN, ContractData, Events, Ledger, Namespace};
+use crate::{
+    deploy::{Deployer, Namespace},
+    Bytes, BytesN, ContractData, Events, Ledger,
+};
+
+#[cfg(doc)]
+use crate::deploy::{CurrentNamespace, Ed25519Namespace};
 
 /// The [Env] type provides access to the environment the contract is executing
 /// within.
@@ -123,16 +129,21 @@ impl Env {
     }
 
     /// Get the contract ID for a contract deployed or that could be deployed.
-    #[doc(hidden)]
     #[inline(always)]
     pub fn contract_id(&self, _namespace: impl Into<Namespace>, _salt: impl Into<Bytes>) -> Events {
         todo!()
     }
 
-    /// Get a [Deployer] for deploying contracts.
+    /// Get a deployer for deploying contracts.
     ///
     /// The namespace given will define the namespace of the contract ID given
     /// to the deployed contract.
+    ///
+    /// To deploy a contract that will be assigned an ID derived from the
+    /// currently executing contract's ID, use [CurrentNamespace].
+    ///
+    /// To deploy a contract that will be assigned an ID derived from an ed25519
+    /// public key, use [Ed25519Namespace].
     #[inline(always)]
     pub fn deployer<D>(&self, namespace: impl Deployer<D>) -> D {
         namespace.deployer(self)
